Added multiple manifest functionality.
Added the ability to change the current installation directory from the options screen.
If the manifest has a Forum URL, it is now loaded in the built-in browser.
Windows can now be resized.
Fixed a bug where -m URL would not work if it was not the first parameter in the parameter list.
Any manifest passed through -m will be added to the manifest list.
Tequila now loads the built-in browser URL from the XML manifest, out of the website node.